pyproject.toml
[project]
name = "hr-recruitment-ai"
version = "0.1.0"
description = "AI-powered recruitment platform for consultants"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.111.0",
    "sqlmodel>=0.0.37",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.6",
    "anthropic>=0.40.0",
    "pdfplumber>=0.11.0",
    "python-pptx>=1.0.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.26.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
